Make default C startup build for X16 and MEGA65
The C64, X16 and MEGA65 are pretty much the same from the starting
point, it seems very plausible that they can share the same header
file, but make different selection attributes to make it look like it
belongs to a specific one (from the same source file), later any minor
changes can be easily code it.

Another advantage is of course that it makes for less duplicate until
it is proved such duplication is really needed (or beneficial).