Define a stack page
Even though the stack page is not needed on Commodore style operating
systems, it is needed for Calypsi to get system stack memory in the
simulator debugger. This is because when loading ELF into the debugger
it only creates simulated memory for memory areas that are part of the
ELF binary.